renamed old config files to config_map
