+++ src/http/config_map/remove.rs
