+++ src/http/mod.rs
@@ -1,3 +1,3 @@
-pub mod config;
+pub mod config_map;
 pub mod token;
 pub mod ping;
